wrap apiRequest for Post/Get/Delete requests with min params
